+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - Issue \#1573: LayerPills: On-click shows layerinfomation for groups, too.
 - Issue \#1578: layer2dVector: Fixed legend creation for WFS layers.
 - Issue \#1590: Vite build - corrected path to mastercode folder in index.html.
+- GroupLayer: Fixed grouped layers, using layerIDsToStyle, were not shown. 
 
 ---